Add files via upload
The dataset includes information on Nike shoes sold across a variety of platforms, including Nike's own website, Amazon, and other retailers. It contains data on the shoe model, colorway, size, price, and the number of reviews and ratings. Additionally, the dataset contains text data on customer reviews, allowing for a detailed analysis of consumer opinions and feedback.

By analyzing this dataset, researchers and analysts can gain insight into which Nike shoes are most popular among consumers, as well as the features and characteristics that customers value most in a shoe. This information could be used to inform future product design and marketing strategies for Nike and other shoe brands.

Furthermore, consumers and shoe enthusiasts can benefit from this dataset to make more informed decisions about purchasing Nike shoes based on the experiences of other customers. The dataset may also be used by retailers to optimize their inventory and marketing strategies based on popular models, sizes, and colorways. Overall, this dataset has the potential to provide valuable insights for both the industry and consumers, shedding light on the factors that contribute to the success of Nike's shoes.